+++ docs/examples/fmu_discipline/sellar/plot_sellar_jacobians.py
@@ -1,25 +1,20 @@
 # Copyright 2021 IRT Saint Exupéry, https://www.irt-saintexupery.com
 #
-# This program is free software; you can redistribute it and/or
-# modify it under the terms of the GNU Lesser General Public
-# License version 3 as published by the Free Software Foundation.
+# This work is licensed under a BSD 0-Clause License.
 #
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-# Lesser General Public License for more details.
+# Permission to use, copy, modify, and/or distribute this software
+# for any purpose with or without fee is hereby granted.
 #
-# You should have received a copy of the GNU Lesser General Public License
-# along with this program; if not, write to the Free Software Foundation,
-# Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
-# Contributors:
-#    INITIAL AUTHORS - initial API and implementation and/or initial documentation
-#        :author: Jorge CAMACHO CASERO
-#    OTHER AUTHORS   - MACROSCOPIC CHANGES
-"""
-Create a discipline for the Sellar problem from a FMU file and computes its Jacobians
-=====================================================================================
-"""
+# THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL
+# WARRANTIES WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED
+# WARRANTIES OF MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL
+# THE AUTHOR BE LIABLE FOR ANY SPECIAL, DIRECT, INDIRECT,
+# OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES WHATSOEVER RESULTING
+# FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN ACTION OF CONTRACT,
+# NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION
+# WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+"""# Create a discipline for the Sellar problem from an FMU file and compute its
+Jacobians"""
 from __future__ import annotations
 
 from gemseo_fmu.problems.disciplines.sellar.sellar_1 import FMUSellar1