+++ src/gemseo_fmu/problems/disciplines/sellar/__init__.py
@@ -23,53 +23,54 @@ In 34th aerospace sciences meeting and exhibit (p. 714).
 
 The MDO problem is written as follows:
 
-.. math::
-
-   \begin{aligned}
-   \text{minimize the objective function }&obj=x_{local}^2 + x_{shared,2}
-   +y_1^2+e^{-y_2} \\
-   \text{with respect to the design variables }&x_{shared},\,x_{local} \\
-   \text{subject to the general constraints }
-   & c_1 \leq 0\\
-   & c_2 \leq 0\\
-   \text{subject to the bound constraints }
-   & -10 \leq x_{shared,1} \leq 10\\
-   & 0 \leq x_{shared,2} \leq 10\\
-   & 0 \leq x_{local} \leq 10.
-   \end{aligned}
+$$
+\begin{aligned}
+\text{minimize the objective function }&obj=x_{\textrm{local}}^2 + x_{\textrm{shared},2}
++y_1^2+e^{-y_2} \\
+\text{with respect to the design variables }&x_{\textrm{shared}},\,x_{\textrm{local}} \\
+\text{subject to the general constraints }
+& c_1 \leq 0\\
+& c_2 \leq 0\\
+\text{subject to the bound constraints }
+& -10 \leq x_{\textrm{shared},1} \leq 10\\
+& 0 \leq x_{\textrm{shared},2} \leq 10\\
+& 0 \leq x_{\textrm{local}} \leq 10.
+\end{aligned}
+$$
 
 where the coupling variables are
 
-.. math::
-
-    \text{Discipline 1: } y_1 = \sqrt{x_{shared,1}^2 + x_{shared,2} +
-     x_{local} - 0.2\,y_2},
+$$
+\text{Discipline 1: } y_1 =
+\sqrt{x_{\textrm{shared},1}^2 + x_{\textrm{shared},2} + x_{\textrm{local}} - 0.2\,y_2},
+$$
 
 and
 
-.. math::
-
-    \text{Discipline 2: }y_2 = |y_1| + x_{shared,1} + x_{shared,2}.
+$$\text{Discipline 2: }y_2 = |y_1| + x_{\textrm{shared},1} + x_{\textrm{shared},2}.$$
 
 and where the general constraints are
 
-.. math::
+$$c_1 = 3.16 - y_1^2$$
 
-   c_1 = 3.16 - y_1^2
+and
 
-   c_2 = y_2 - 24
+$$c_2 = y_2 - 24.$$
 
 This package implements three disciplines
 to compute the different coupling variables, constraints and objective:
 
-- :class:`.FMUSellar1`:
-  this :class:`.MDODiscipline` computes :math:`y_1`
-  from :math:`y_2`, :math:`x_{shared,1}`, :math:`x_{shared,2}` and :math:`x_{local}`.
-- :class:`.FMUSellar2`:
-  this :class:`.MDODiscipline` computes :math:`y_2`
-  from :math:`y_1`, :math:`x_{shared,1}` and :math:`x_{shared,2}`.
-- :class:`.FMUSellarSystem`:
-  this :class:`.MDODiscipline` computes both objective and constraints
-  from :math:`y_1`, :math:`y_2`, :math:`x_{local}` and :math:`x_{shared,2}`.
+- [FMUSellar1][gemseo_fmu.problems.disciplines.sellar.sellar_1.FMUSellar1]:
+  this [FMUDiscipline][gemseo_fmu.disciplines.fmu_discipline.FMUDiscipline]
+  computes $y_1$
+  from $y_2$, $x_{\textrm{shared},1}$, $x_{\textrm{shared},2}$ and $x_{\textrm{local}}$.
+- [FMUSellar2][gemseo_fmu.problems.disciplines.sellar.sellar_2.FMUSellar2]:
+  this [FMUDiscipline][gemseo_fmu.disciplines.fmu_discipline.FMUDiscipline]
+  computes $y_2$
+  from $y_1$, $x_{\textrm{shared},1}$ and $x_{\textrm{shared},2}$.
+- [FMUSellarSystem][gemseo_fmu.problems.disciplines.sellar.sellar_system.FMUSellarSystem]:
+  this [FMUDiscipline][gemseo_fmu.disciplines.fmu_discipline.FMUDiscipline]
+  computes both objective and constraints
+  from $y_1$, $y_2$, $x_{\textrm{local}}$ and $x_{\textrm{shared},2}$.
 """
 from __future__ import annotations